Turned back on cython parsers now that the parsers are more decoupled from schema's that use them.  Seems to be 2x faster.
Also added a schema change to the meta store to represent the date
of last fetched item.  This enables proper restarting from files.
The issue is that if I call the API today and it has data from
year 2000 to 5 days ago, when I restart it will try to call the
API again.  This slows restarts on days where it is parsing.
It's also better to have a mechanism that is not reliant on
MySQL for restarting in case of the database gets dropped.
The fetch date is embedded in the file path since it is not in the
xml file itself.